"_tkinter" runtime hook fixed.
This commit corrects the "_tkinter" runtime hook in accordance with changes to
the "_tkinter" module hook. For orthogonality with other hooks, Tcl and Tk data
directories are now frozen into the top-level directory of their frozen archive
rather than embedded in a "_MEI" subdirectory. For safety, the existence of such
directories is now guaranteed with exceptions.